Get the shortest path from xyStart to xyEnd
The function creates a starting Node -> source,
the xyStart. If the end is not the same as start
then we use an while loop.

In there we dynamically create new paths/nodes
corresponding to the legal moves of the knight.
We check the position of the currently inspected
Node that is shifted from out queue. The queue
implements BFS.