Add a logging message
In the Project assignment, it wass recommended to
check if some path/xy was already checked so we
don't go into infinite loop.

So lets say we start at 0,0 -> we shouldn't check
this coordinates again.

But in my code - we always check the current position
- so if this is not the final position, it's "okay"
to check again in the child's neighbours. Cause if
it is the final, then we don't even get to the children
basically.